daa_sim/requirements-dev.txt
PyYAML~=6.0
uvicorn[standard]
fastapi
qiskit>=1.2.2
qiskit-aer>=0.15.1
qiskit_ibm_runtime>=0.30.0
qiskit_qasm3_import
numpy>=1.26,<2
sympy>=1.13
urllib3>=2.2.3
boto3
pyjwt

# Style and Lint: Following Qiskit 1.2
black[jupyter]~=24.1
ruff==0.0.267
pylint==3.2.3
